Changed network latency to a decimal value

+++ lib/services/nt_connection.dart
@@ -116,7 +116,7 @@ class NTConnection {
     }
   }
 
-  Stream<int> latencyStream() {
+  Stream<double> latencyStream() {
     return nt4Client.latencyStream();
   }
